Split apart and linting passes
