+++ pyop2/types/access.py
@@ -1,0 +1,37 @@
+import enum
+
+
+class Access(enum.IntEnum):
+    READ = 1
+    WRITE = 2
+    RW = 3
+    INC = 4
+    MIN = 5
+    MAX = 6
+
+
+READ = Access.READ
+"""The :class:`Global`, :class:`Dat`, or :class:`Mat` is accessed read-only."""
+
+WRITE = Access.WRITE
+"""The  :class:`Global`, :class:`Dat`, or :class:`Mat` is accessed write-only,
+and OP2 is not required to handle write conflicts."""
+
+RW = Access.RW
+"""The  :class:`Global`, :class:`Dat`, or :class:`Mat` is accessed for reading
+and writing, and OP2 is not required to handle write conflicts."""
+
+INC = Access.INC
+"""The kernel computes increments to be summed onto a :class:`Global`,
+:class:`Dat`, or :class:`Mat`. OP2 is responsible for managing the write
+conflicts caused."""
+
+MIN = Access.MIN
+"""The kernel contributes to a reduction into a :class:`Global` using a ``min``
+operation. OP2 is responsible for reducing over the different kernel
+invocations."""
+
+MAX = Access.MAX
+"""The kernel contributes to a reduction into a :class:`Global` using a ``max``
+operation. OP2 is responsible for reducing over the different kernel
+invocations."""
